+++ lib/pmdtester/builders/diff_builder.rb
@@ -11,12 +11,12 @@ module PmdTester
     def build(base_report_filename, patch_report_filename, base_info, patch_info, filter_set = nil)
       report_diffs = ReportDiff.new
       base_details, patch_details = report_diffs.calculate_details(base_info, patch_info)
-      base_violations, base_errors = parse_pmd_report(base_report_filename, BASE,
-                                                      base_details.working_dir, filter_set)
-      patch_violations, patch_errors = parse_pmd_report(patch_report_filename, PATCH,
-                                                        patch_details.working_dir)
-      report_diffs.calculate_violations(base_violations, patch_violations)
-      report_diffs.calculate_errors(base_errors, patch_errors)
+      base_report = parse_pmd_report(base_report_filename, BASE, base_details.working_dir,
+                                     filter_set)
+      patch_report = parse_pmd_report(patch_report_filename, PATCH, patch_details.working_dir)
+      report_diffs.calculate_violations(base_report.violations, patch_report.violations)
+      report_diffs.calculate_errors(base_report.errors, patch_report.errors)
+      report_diffs.calculate_configerrors(base_report.configerrors, patch_report.configerrors)
 
       report_diffs
     end
@@ -25,7 +25,7 @@ module PmdTester
       doc = PmdReportDocument.new(branch, working_dir, filter_set)
       parser = Nokogiri::XML::SAX::Parser.new(doc)
       parser.parse_file(report_filename) unless report_filename.nil?
-      [doc.violations, doc.errors]
+      doc
     end
   end
 end