+++ test/test_diff_builder.rb
@@ -19,10 +19,10 @@ class TestDiffBuilder < Test::Unit::TestCase
     diffs_report = diff_builder.build(base_report_path, patch_report_path,
                                       BASE_REPORT_INFO_PATH, PATCH_REPORT_INFO_PATH)
     violation_diffs = diffs_report.violation_diffs
-    error_diffs = diffs_report.error_diffs
     keys = violation_diffs.keys
 
-    assert_empty(error_diffs)
+    assert_empty(diffs_report.error_diffs)
+    assert_empty(diffs_report.configerrors_diffs)
     assert_equal(5, diffs_report.base_violations_size)
     assert_equal(8, diffs_report.patch_violations_size)
     assert_equal(7, diffs_report.violation_diffs_size)
@@ -45,11 +45,11 @@ class TestDiffBuilder < Test::Unit::TestCase
     patch_report_path = 'test/resources/diff_builder/test_error_diffs_patch.xml'
     diffs_report = diff_builder.build(base_report_path, patch_report_path,
                                       BASE_REPORT_INFO_PATH, PATCH_REPORT_INFO_PATH)
-    violation_diffs = diffs_report.violation_diffs
     error_diffs = diffs_report.error_diffs
     keys = error_diffs.keys
 
-    assert_empty(violation_diffs)
+    assert_empty(diffs_report.violation_diffs)
+    assert_empty(diffs_report.configerrors_diffs)
     assert_equal(4, diffs_report.base_errors_size)
     assert_equal(3, diffs_report.patch_errors_size)
     assert_equal(5, diffs_report.error_diffs_size)
@@ -61,6 +61,31 @@ class TestDiffBuilder < Test::Unit::TestCase
     assert_equal('Both2.java', keys[1])
     assert_equal(2, error_diffs['Both2.java'].size)
     assert_equal('Patch1.java', keys[2])
+    assert_equal(true, diffs_report.introduce_new_errors?)
+  end
+
+  def test_configerrors_diffs
+    diff_builder = DiffBuilder.new
+    base_report_path = 'test/resources/diff_builder/test_configerrors_diffs_base.xml'
+    patch_report_path = 'test/resources/diff_builder/test_configerrors_diffs_patch.xml'
+    diffs_report = diff_builder.build(base_report_path, patch_report_path,
+                                      BASE_REPORT_INFO_PATH, PATCH_REPORT_INFO_PATH)
+    configerrors_diffs = diffs_report.configerrors_diffs
+    keys = configerrors_diffs.keys
+
+    assert_empty(diffs_report.violation_diffs)
+    assert_empty(diffs_report.error_diffs)
+    assert_equal(4, diffs_report.base_configerrors_size)
+    assert_equal(3, diffs_report.patch_configerrors_size)
+    assert_equal(5, diffs_report.configerrors_diffs_size)
+    assert_equal(3, diffs_report.removed_configerrors_size)
+    assert_equal(2, diffs_report.new_configerrors_size)
+    assert_equal(3, keys.size)
+    assert_equal('RuleBase1', keys[0])
+    assert_equal(2, configerrors_diffs['RuleBase1'].size)
+    assert_equal('RuleBoth2', keys[1])
+    assert_equal(2, configerrors_diffs['RuleBoth2'].size)
+    assert_equal('RulePatch1', keys[2])
     assert_equal(true, diffs_report.introduce_new_errors?)
   end