+++ History.md
@@ -12,6 +12,7 @@
 *   [#46](https://github.com/pmd/pmd-regression-tester/issues/46): Support multithreaded execution of PMD
 *   [#49](https://github.com/pmd/pmd-regression-tester/issues/49): Support comparing two error stacktraces
 *   [#50](https://github.com/pmd/pmd-regression-tester/issues/50): Differences due to different locale settings
+*   [#57](https://github.com/pmd/pmd-regression-tester/issues/57): Support configuration errors in the report
 
 ## External Contributions